Moved database.ts to /services to centralise all DB-related functionality
